Added: Controller2DPhysics. This script does four things: first it allows you to use physics terms such as velocity and acceleration rather than the Controller2D move command. Second it allows scripts to segregate their velocity inputs i.e. a moving platform can move the player up and down with out interfering with the player's ability to jump. Third it allows a script to lock the player into a specific velocity for a short period of time i.e. basis for a knock back or stun script. Finally it establishes a motion hierarchy: collisions overrides motion locks overrides physics.
Movement_Experimental is an experimental jumping script that currently allows variable jump heights by manipulating the gravity acceleration based on a timer. It is a start but the differences in acceleration are visibly noticeable and makes for odd jerky jumps. I will experiment with the script, possibly by modifying both the initial velocity and the acceleration to create a smoother variable jump.

Currently the code is uncommented. That will be fixed at a later time